+++ NewHorizons/Builder/Volumes/PriorityVolumeBuilder.cs
@@ -1,4 +1,4 @@
-using NewHorizons.External.Modules.Volumes;
+using NewHorizons.External.Modules.Volumes.VolumeInfos;
 using UnityEngine;
 
 namespace NewHorizons.Builder.Volumes
