+++ NewHorizons/Builder/Volumes/VisorEffects/VisorFrostEffectVolumeBuilder.cs
@@ -1,4 +1,5 @@
 using NewHorizons.External.Modules.Volumes;
+using NewHorizons.External.Modules.Volumes.VolumeInfos;
 using UnityEngine;
 
 namespace NewHorizons.Builder.Volumes.VisorEffects
