+++ NewHorizons/External/Modules/SerializableEnums/DeathType.cs
@@ -2,7 +2,7 @@ using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 using System.Runtime.Serialization;
 
-namespace NewHorizons.External.Modules.Volumes
+namespace NewHorizons.External.Modules.SerializableEnums
 {
     [JsonConverter(typeof(StringEnumConverter))]
     public enum DeathType
