+++ NewHorizons/External/Modules/Volumes/VolumeInfos/LoadCreditsVolumeInfo.cs
@@ -1,7 +1,8 @@
+using NewHorizons.External.Modules.SerializableEnums;
 using Newtonsoft.Json;
 using System.ComponentModel;
 
-namespace NewHorizons.External.Modules.Volumes
+namespace NewHorizons.External.Modules.Volumes.VolumeInfos
 {
     [JsonObject]
     public class LoadCreditsVolumeInfo : VolumeInfo
