+++ NewHorizons/External/Modules/Volumes/VolumeInfos/OxygenVolumeInfo.cs
@@ -1,7 +1,7 @@
 using Newtonsoft.Json;
 using System.ComponentModel;
 
-namespace NewHorizons.External.Modules.Volumes
+namespace NewHorizons.External.Modules.Volumes.VolumeInfos
 {
     [JsonObject]
     public class OxygenVolumeInfo : VolumeInfo
