+++ NewHorizons/External/Modules/Volumes/VolumeInfos/VolumeInfo.cs
@@ -1,0 +1,18 @@
+using NewHorizons.External.Modules.Audio;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+using System.ComponentModel;
+using System.ComponentModel.DataAnnotations;
+using System.Runtime.Serialization;
+
+namespace NewHorizons.External.Modules.Volumes.VolumeInfos
+{
+    [JsonObject]
+    public class VolumeInfo : GeneralPointPropInfo
+    {
+        /// <summary>
+        /// The radius of this volume.
+        /// </summary>
+        [DefaultValue(1f)] public float radius = 1f;
+    }
+}
